% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MUPPITS1bloodDesign}
\alias{MUPPITS1bloodDesign}
\title{MUPPITS1bloodDesign}
\format{
An object of class \code{data.frame} with 374 rows and 128 columns.
}
\usage{
MUPPITS1bloodDesign
}
\description{
MUPPITS1bloodDesign - MUPPITS1 Blood design object
}
\keyword{datasets}
