% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MUPPITS1bloodModules}
\alias{MUPPITS1bloodModules}
\title{MUPPITS1bloodModules}
\format{
An object of class \code{data.frame} with 13374 rows and 4 columns.
}
\usage{
MUPPITS1bloodModules
}
\description{
MUPPITS1bloodModules - modules made from blood/tempus data
}
\keyword{datasets}
