% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MUPPITS1bloodVoom}
\alias{MUPPITS1bloodVoom}
\title{MUPPITS1bloodVoom}
\format{
An object of class \code{EList} with 13316 rows and 387 columns.
}
\usage{
MUPPITS1bloodVoom
}
\description{
MUPPITS1bloodVoom - MUPPITS1 Blood Voom object
}
\keyword{datasets}
