% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MUPPITS1nasalDesign}
\alias{MUPPITS1nasalDesign}
\title{MUPPITS1nasalDesign}
\format{
An object of class \code{data.frame} with 387 rows and 121 columns.
}
\usage{
MUPPITS1nasalDesign
}
\description{
MUPPITS1nasalDesign - MUPPITS1 Nasal design object
}
\keyword{datasets}
