% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MUPPITS1nasalModules}
\alias{MUPPITS1nasalModules}
\title{MUPPITS1nasalModules}
\format{
An object of class \code{data.frame} with 14220 rows and 4 columns.
}
\usage{
MUPPITS1nasalModules
}
\description{
MUPPITS1nasalModules - modules made from nasal transcriptome data
}
\keyword{datasets}
