% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MUPPITS1nasalVoom}
\alias{MUPPITS1nasalVoom}
\title{MUPPITS1nasalVoom}
\format{
An object of class \code{EList} with 13672 rows and 374 columns.
}
\usage{
MUPPITS1nasalVoom
}
\description{
MUPPITS1nasalVoom - MUPPITS1 Nasal Voom object
}
\keyword{datasets}
